Ensure target is up to date in android run config handler.
Due to the way blaze run configuration editors update the underlying
configuration state, it's possible for the configuration referenced in
android run config handlers to be out of date. This can cause tricky
side-effects such as incorrect build target being build during test.

This CL ensures the target is up to date when a new runner is created
from the handler by updating it from the execution environment.

PiperOrigin-RevId: 254265159